refactor(client): move admin client
